Huge interface-breaking commit. Like, really huge.
General idea: we are introducing an `xml2human.py` script that accepts a
general input and prints a pretty human-readable version of the input.
In order to make it easier to use, we wanted to detect the format of the
input file automatically.

The problem is that we were previously subclassing `AbstractParser` directly
whenever we wanted to listen to parse events, and that made it impossible to
delegate work to other parsers. So this patch has moved this handling behavior
to `parser_wrappers.InputHandler`; and this is the class that should be
subclassed from now on.  The correct parser will be automatically detected when
calling `parser_wrappers.parse`.