Fixed backtrace handling.
Fixed cases where `--debug` would disable the printing of backtraces and added an
exception type MWEToolkitInputError for user input errors (by default, most of
the stack trace is omitted, since it is not a bug in the code).

Now the debug mode only forces printing the whole backtrace on any error (and
also prints a backtrace on warnings).