Fun side-project: Added support for compressed files.
Now, something like `$MWETOOLKIT/head.py corpus.xml.gz` should work
seamlessly.  Supported formats are GZIP, ZIP and BZ2 (but due to restrictions
in python's `bz2` library, BZ2 will only work in Python 3.3+)

TAR is not supported, because I don't like it:  it is not a compression
method and it does not have a pretty magic header like all the other cool kids.