Making `uniq.py` deterministic.
Lists of features/occurrences/bigrams/... have been replaced either by sets
(when the order didn't matter, as in "list of ngrams") or by OrderedDicts
(when the ordered was important). Features, in particular, are now stored in
`FeatureSet` objects (which use OrderedDicts internally).

-- To represent sets of features, replacing `list`s by `FeatureSet` objects.
-- Now avoiding `set` or using `sorted` afterwards.
-- Not using `shelve` anymore, because Mac systems have a broken
implementation.  Users can enable virtual memory if they really need to process
huge lists of patterns or MWE candidates.  (Corpora are never loaded fully in
memory).
-- Now `sort.py` is ascending by default.
-- Refactored `uniq.py`. Removed buggy `--retokenize` option.
-- Updated lots of tests. Now all tests involving `uniq.py` should pass.

In the future, we may deprecate the Feature class, or at least replace it with
a namedtuple, as that's essentially a key-value pair, and FeatureSet objects
will actually store this pair as a subclass of dict (thus dismembering the
feature object internally anyway).